WIP Use Evaluator.findResolver() to avoid hardcoded resolver.
I don't know if this is right; would be better to use a parameter and not hard-code in the Evaluator.  But that requires having the Evaluator everywhere we need to resolve, and that's not currently true at the Runtime level.

WARNING This was rebased from a very old branch, some conflicts may be misaddressed.